requirements.txt
aniso8601==3.0.2
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
croniter==0.3.25
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Cors==3.0.6
Flask-Dance==1.0.0
Flask-Principal==0.4.0
flask-restplus==0.11.0
Flask-RQ2==18.0
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy==1.3
MarkupSafe==1.0
oauthlib==2.1.0
PyJWT==1.6.4
pymongo==3.7.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rq==0.12.0
rq-scheduler==0.8.3
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.2.1
